zip files before downloading
